Change the timezones to the match the city portion of the canonical IANA Zone IDs
This renames the following Zone IDs:

* Beijing -> Shangai
* Boulder -> Denver
* GST -> Los_Angeles
* India -> Kolkata
* Israel -> Jerusalem
* Japan -> Tokyo

The new names are the *city* portion of the canonical Zone ID for each zone. (See https://nodatime.org/TimeZones for a complete list of canonical Zone IDs from IANA 2023c.)

The most important change is “GST -> Los_Angeles”. In the code, “GST” was taken to mean “Google Standard Time”, but it already means something else: “Gulf Standard Time” (https://www.timeanddate.com/time/zones/gst).

In addition:

 * The India zoneinfo file wasn’t YOINKed; now it is, under the name Kolkata
 * The Melbourne zoneinfo file *was* YOINKed; it was deleted in favor of an already-existing un-YOINKed Sydney file
 * The Singapore zoneinfo file wasn’t YOINKed; it was deleted

(Melbourne is in the same timezone as Sydney. Singapore is in the same timezone as Shanghai. But note that “same timezone” is misleading, since *when* the standard time/daylight savings time shift happens is a political question.)

Finally, the offsets in the docs were a mixture of standard time and daylight savings time. Now both offsets are given. The first is always the standard time offset. The second, if applicable, is the daylight savings offset. Among other things, that clarifies which timezones do and do not have daylight savings.