Paginate by writing data to pager via stdin
On Windows, `more` is unable to open the file generated by `mkstemp()`
due to not recognizing the first argument as a path, but as a parameter.
In contrast to *nix, arguments on Windows are, by convention, prefixed
with `/`, which `mkntcmdline()` happily transforms into path-like strings.

This commit sidesteps this issue entirely by writing data to the pager
in question via stdin, dumping text to the console as a last resort.