vcscanf: parse floating point numbers
The vcscanf function attempts to recognize the floating point number
and then uses the strtod function to calculate its numeric value.
It begins by recognizing an optional plus or minus sign, followed by
either hexadecimal notation or NaN and infinity constants.
Then it looks for an optional integer part, a point or comma,
a fractional part, and an optional exponent symbol.
If it finds an exponent, it recognizes the optional sign and digits.

Every time a character is accepted, it is also buffered.
The accumulated buffer is passed to strtod when recognition is finished
and the output is assigned to the output pointer of appropriate length.
This only happens if not in discard mode.
Finally, the buffer is freed and its state is reset.

References:

https://en.cppreference.com/w/c/string/byte/strtof