Fix zipos deadlock/segfault
Locks `g_fds` for /zip `close()` calls. Moves lock to the top level in
`__zipos_postdup`. The latter needs review for signal blocking
correctness; it is no more wrong than it was before on that front, but
I'm not at all confident that it was ever right.

Added a defensive assert in `__zipos_read`.