sanity-check set-id interpreter script paths
The check is fairly stringent; we do it in a constructor, and we make
sure that there actually is an open fd higher than 2 that matches the
passed path.

There is an argument to be made for trapping in the constructor here,
and there is also an argument to be made for failing open like we did
previously. We pick a middle ground: GetProgramExecutableName returns
empty string. Logic being that ensuring the sanity of whatever set-id
script mechanism user is running is above our pay grade, but the path
TOCTOU between loader and binary is not.