Remove constructor, do not verify fd with fcntl
By the logic of the previous commit, it's not our job to verify that the
user's whacked out set-id configuration is sane, only that the path they
passed does not refer to a different file now than it did when ape found
it. This is generally true of the files under /dev/fd in our case.

So we just check that the file is in /dev/fd/, that it starts with not a
'.', and that it does not contain a '/'.

Since we don't care whether there actually is an fd by that name, we can
do the check when called and save a constructor.