Skip tests when they will fail
GetProgramExecutableName does not work right on assimilated OpenBSD or
non-silicon XNU binaries. If we are on one of those platforms, look for
ape magic in ourselves and quit if none is found.

If `argv[0]` is not reliable for the test invocation itself (e.g.
because the test is run out of `PATH`), the test assert-fails.