Aggressively simplify MODE=tiny EU time zones
I took one canonical IANA zone ID from each of the different colored
regions in this article, except those that do not observe DST and do
not at present contain a Google office:

<https://en.wikipedia.org/wiki/Time_in_Europe>

As to which canonical ID to use, this was somewhat arbitrary. Brussels
was obvious, as the de facto capital of the EU. For the rest, I mostly
just went with lexicographic ordering of the most recognizable options.