loader.c uses AT_FLAGS to communicate literally
Now the `ape - prog args...` form communicates to the loaded binary that
argv[0] is preserved. This is done via either editing the previous flags
passed in auxv, or adding a new AT_FLAGS value at the end if one was not
found.

I decided to always insert an AT_FLAGS entry, although another option is
to only insert one if literally is true. This would change the value for
`n` to `... + 2 * (!flags && literally) + ...` and add another branch to
the if statement. I guessed that the extra branch was not worth saving 2
words on the stack sometimes.